Embedded album art (ticket #2)
Any embedded pictures present in the flac file are transferred to the
mp3 file. Will only update picture tag if album art is changed.

FreezeThaw is used to stringify and then compare the contents of the
picture tags. Since FreezeThaw is not in core in ActiveState Perl for
MSWin (as of version 5.16.1) the module is included in the library.

This code has been refactored to blend in better with the rest of the
tagging activities in the script (as compared to my earlier version).