revert temporarily to "copy non-flac files"
Added option "--copyfiles" (Copy non-flac files to dest
directories)

When this option is selected, non-flac files present in the source
directories are copied to the target directories (which, for some, is
another way of handling album
art). Copying only takes place when the source and target md5
hashes don't match.