refactor path_and_convert sub (step 1/5)
In a first step towards progress reporting during transcoding,
refactor code for the 'path_and_conversion' sub. The checking of
flags that determine whether a file should be transcoded is
lifted into this sub from 'transcode_file'. The latter sub
therefore loses a wrapping if-statement and consequently also
loses an indentation step, which makes the changeset look
much messier than it actually is.

The sub 'convert_file' is no longer used.