buffer message output (step 2/5)
In this second step towards progress reporting the messages from
transcoding and tagging are not display immediately on screen, but
saved to an array variable and displayed once the respective forked
child exits. Without this step, there is no real control over the
order by with the messages appear when using several parallel
processes, and progress reporting would look a complete mess.

A callback subroutine is used to output the messages. It is
called via the method 'run_on_finish' from the module
Parallel::Forkmanager.