progress report after each transcode (step 4/5)
Display total flac file size to convert before transcoding
begins (in units of MB), as well as a progress report after
each transcoded file (as in "Processed 3/1267 files."). This
is accomplished by extending the callback code that runs
every time a child exits.

The timing and estimated finish time of the transcoding job is
implemented by extending this callback code further in the
next commit, step 5/5.