estimate time remaining (step 5/5)
Present timing of each transcode to 0.1 s resolution and also,
after 20 s of transcoding, display the estimated time remaining
for the whole job. The latter time is extapolated from the
total file size converted so far and from the time elapsed since
transcoding begun.

The estimed time remaining is updated each time a file is
transcoded. The accuracy of the estimation will successively
improve during the job as more and more files are transcoded.

This is the fifth and final changeset for the step-by-step
implementation of progress reporting.